+++ eTaxi/eTaxi.Application/DTOs/Vehicle/VehicleDto.cs
@@ -14,7 +14,7 @@
         public int? CurrentLocationId { get; set; }
         public string Color { get; set; }
         public string Brand { get; set; }
-        public int PricePerKm { get; set; }
+        public double PricePerKm { get; set; }
         public int? UserDriverId { get; set; }
         public string ImageUrl { get; set; }
         public Domain.VehicleType Type { get; set; }
